// enhanced/meta/rm_attribute_meta.generated.ts
export interface RmAttributeMetaRow {
  name: string;
  typeName: string;
  min: number;
  max: number | null;
  mandatory: boolean;
  documentation?: string;
  polymorphic: boolean;
}

export const RM_META_SCHEMA_NAME = "rm";
export const RM_META_RM_RELEASE = "1.2.0";
export const RM_META_BMM_VERSION = "2.4";
export const RM_META_SOURCES: readonly string[] = [
  "https://raw.githubusercontent.com/sebastian-iancu/code-generator/master/code/BMM-JSON/openehr_base_1.3.0.bmm.json",
  "https://raw.githubusercontent.com/sebastian-iancu/code-generator/master/code/BMM-JSON/openehr_rm_1.2.0.bmm.json"
];
export const RM_META_GENERATED_AT = "2026-07-21T07:02:42.058Z";